requirements.txt
pandas==1.3.5
numpy==1.21.6
matplotlib==3.5.1
seaborn==0.11.0
tqdm==4.64.1
torch==1.13.1
Pillow==9.1.1
torchvision==0.14.1
opencv-python==4.6.0.66
tensorflow==2.9.2
keras==2.9.0
scikit-image==0.19.3
scipy==1.7.3